Rename ContestOutdated to ContestUnusedDec
Add Healthy module for contest mutations and ContestUnusedDec module for
corresponding mutations.
