+++ hydra-cluster/config/devnet/genesis-shelley.json
@@ -22,7 +22,7 @@
         "keyDeposit": 0,
         "maxBlockBodySize": 65536,
         "maxBlockHeaderSize": 1100,
-        "maxTxSize": 17900,
+        "maxTxSize": 18000,
         "minFeeA": 44,
         "minFeeB": 155381,
         "minPoolCost": 0,
