+++ hydra-tx/test/Hydra/Tx/Contract/ContractSpec.hs
@@ -39,7 +39,8 @@ import Hydra.Tx.Contract.Close.CloseUnused (genCloseCurrentMutation, healthyClos
 import Hydra.Tx.Contract.Close.CloseUsed (genCloseOutdatedMutation, healthyCloseOutdatedTx)
 import Hydra.Tx.Contract.CollectCom (genCollectComMutation, healthyCollectComTx)
 import Hydra.Tx.Contract.Commit (genCommitMutation, healthyCommitTx)
-import Hydra.Tx.Contract.Contest.ContestCurrent (genContestMutation, healthyContestTx)
+import Hydra.Tx.Contract.Contest.ContestCurrent (genContestMutation)
+import Hydra.Tx.Contract.Contest.Healthy (healthyContestTx)
 import Hydra.Tx.Contract.Decrement (genDecrementMutation, healthyDecrementTx)
 import Hydra.Tx.Contract.Deposit (healthyDepositTx)
 import Hydra.Tx.Contract.FanOut (genFanoutMutation, healthyFanoutTx)
